copdb/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
boto3==1.4.7
botocore==1.7.43
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.9
django-appconf==1.0.4
django-celery-email==3.0.0
django-cors-headers==3.2.1
django-countries==6.1.2
django-crispy-forms==1.9.1
django-database-size==1.0.2
django-db-logger==0.1.9
django-localflavor==3.0.1
django-storages==1.9.1
django-templated-mail==1.1.1
django-videofield==0.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.12
elementtree===1.2.7-20070827-preview
fcm-django==0.3.4
gender-guesser==0.4.0
geographiclib==1.50
geoip2==3.0.0
geopy==1.21.0
gevent==1.3.7
greenlet==0.4.15
html5lib==0.90
idna==2.9
importlib-metadata==1.6.0
jmespath==0.9.3
kombu==4.6.8
libsass==0.15.1
lxml==4.2.5
Markdown==2.6.9
MarkupSafe==0.23
maxminddb==1.5.2
meson==0.51.0
mysqlclient==1.3.14
ninja==1.9.0
numpy==1.19.1
olefile==0.46
packaging==19.2
pandas==1.1.1
pbr==2.0.0
Pillow==5.4.0
psutil==4.3.1
psycopg2==2.7.3.1
psycopg2-binary==2.8.5
pycairo==1.15.3
pycrypto==2.6.1
pycurl==7.43.0.5
pydot==1.2.3
pyfcm==1.4.7
pyldap==2.4.28
pyparsing==2.4.5
python-dateutil==2.8.1
python-http-client==3.2.7
python-Levenshtein==0.12.0
python-openid==2.2.5
python-stdnum==1.13
pytz==2020.1
PyYAML==3.13
reportlab==3.5.31
requests==2.23.0
s3transfer==0.1.11
scandir==1.5
scikit-build==0.10.0
sendgrid==6.3.1
simplejson==3.11.1
six==1.15.0
sqlparse==0.3.0
stevedore==1.30.0
stripe==2.48.0
supervisor==4.2.0
timezonefinder==4.3.1
urllib3==1.25.9
vine==1.3.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
zipp==3.1.0
django-rest-multiple-models
